test: add comprehensive validation tests for file initialization
- Add test for invalid JSON in sessionStorage (parse error handling)
- Add test for non-array JSON data validation
- Add test for files with missing required properties (name, type, size)
- Add test for invalid dataUrl validation (non-string type)
- Add test for invalid file data structure (non-object, null values)
- Extract handleShareTargetMessage to ShareTarget.utils.ts for better module organization
- Fix react-refresh/only-export-components ESLint warning

These tests improve coverage for error handling paths in file initialization
and validation logic, targeting the remaining ~5.7% coverage gap to reach
the 81.42% Codecov requirement.
